#10 jboss_confignode in puppet_x
