+++ tests/clientry.pp
@@ -1,0 +1,8 @@
+include jboss
+
+jboss::clientry { '/subsystem=messaging/hornetq-server=default':
+  ensure     => 'present',
+  properties => {
+    'security-enabled' => false,
+  }
+}
